Create TypoScriptFrontendController in backend context
Under certain circumstances it is neccessary to have a TypoScriptFrontendController available to avoid errors. For example the backend module of system extension redirects creates frontend links in backend context. This executes the urlProcessors hook of jumpurl but fails because there is no TypoScriptFrontendController. Method getTypoScriptFrontendController in class TYPO3\CMS\Frontend\Typolink\AbstractTypolinkBuilder provides some kind of workaround. This change is basically just copy and paste of this workaround.